Bit of Figurine Work
Not super complete, was just poking around to figure a few things out

+++ include/d/d_snap.h
@@ -320,7 +320,7 @@ public:
 };
 
 int dSnap_PhotoIndex2TableIndex(int);
-u8 dSnap_GetFigRoomId(int);
+int dSnap_GetFigRoomId(int);
 void dSnap_MatDl();
 void dSnap_AlphaClearDL();
 void dSnap_Create();
